# node/middlewares/db/models/Movie.js
const mongoose = require('mongoose');

const MovieSchema = new mongoose.Schema({
    title: {
        type: String,
        required: true,
    },
    description: {
        type: String,
        required: true,
    },
    poster: {
        type: String,
        required: true,
    },
    year: {
        type: Number,
        required: true,
    },
    genre: [
        {
            type: String,
            required: true,
        },
    ],
});

const Movie = mongoose.model('Movie', MovieSchema);

module.exports = Movie;